requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.6.0
cycler==0.11.0
Django==4.0.3
django-cors-headers==3.13.0
django-multi-form-view==2.0.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.31.2
future==0.18.2
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
html5lib==1.1
httplib2==0.20.4
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.1
numpy==1.22.3
oauth2client==4.1.3
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive2==1.10.1
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
spyne==2.14.0
sqlparse==0.4.2
svglib==1.2.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
xlwt==1.3.0
